Public test 3 Rename EPG files and classes. Dual skin screens names used to maintain skin compatibility Fix crash when pressing Info on Channel Selection Fix EPGSearch crash when show orbital is on
